app/requirements.txt
# YTMusic App - Python Dependencies
# Install with: pip install -r requirements.txt

pywebview>=5.0.0
ytmusicapi>=1.0.0

